Commented out the string replacement template - it isn't needed, and it collides with the character replacement template (they had exactly the same match pattern - no good).
Reformulated the match for the character replacement target so that it will only target the relevant subset of elements, namely those *with* text content that do *not* contain em or span elements, and are within the body element. This will ensure that em and span markup within p elements is not destroyed (and the element containing the bad characters do not contain other markup), and also that empty elements will be deleted. (by another template).

This was a very useful learning experience:)